One sln file for Doc tools, fixes up standalone overview files to have a header
