+++ src/docs/JavaDocToMarkdownConverter/Formatters/DocTypeReplacer.cs
