+++ src/docs/JavaDocToMarkdownConverter/Formatters/NamedAnchorLinkReplacer.cs
