+++ src/docs/JavaDocToMarkdownConverter/Formatters/PatternReplacer.cs
