+++ src/docs/JavaDocToMarkdownConverter/Properties/AssemblyInfo.cs
