+++ src/docs/JavaDocToMarkdownConverter/StringExtensions.cs
