+++ src/docs/LuceneDocsPlugins/LuceneDocsPlugins.csproj
