+++ src/docs/LuceneDocsPlugins/LuceneRendererPartProvider.cs
