+++ src/docs/LuceneDocsPlugins/LuceneTokenRendererPart.cs
