+++ src/docs/LuceneDocsPlugins/Properties/AssemblyInfo.cs
