+++ src/docs/LuceneDocsPlugins/app.config
