+++ src/docs/LuceneDocsPlugins/packages.config
