+++ websites/apidocs/lucenetemplate/conceptual.html.primary.tmpl
@@ -15,23 +15,15 @@
  limitations under the License.
 }}
 
-<nav id="autocollapse" class="navbar ng-scope" role="navigation">
-  <div class="container">
-    <div class="navbar-header">
-      <button type="button" class="navbar-toggle" data-toggle="collapse" data-target="#navbar">
-        <span class="sr-only">Toggle navigation</span>
-        <span class="icon-bar"></span>
-        <span class="icon-bar"></span>
-        <span class="icon-bar"></span>
-      </button>
-      {{>partials/logo}}
-    </div>
-    <div class="collapse navbar-collapse" id="navbar">
-      <form class="navbar-form navbar-right" role="search" id="search">
-        <div class="form-group">
-          <input type="text" class="form-control" id="search-query" placeholder="Search" autocomplete="off">
-        </div>
-      </form>
-    </div>
-  </div>
-</nav>
+{{!master(layout/_master.tmpl)}}
+
+{{#rawTitle}}
+  {{{rawTitle}}}
+{{/rawTitle}}
+{{^rawTitle}}
+  {{#title}}
+    <h1 id="{{id}}" data-uid="{{uid}}" class="text-break">{{title}}</h1>
+  {{/title}}
+{{/rawTitle}}
+
+{{{conceptual}}}
